Fix paths after moving into api directory
