+++ manageiq-operator/api/v1alpha1/helpers/cr_migration/20210503163000_initial_migration.go
@@ -1,7 +1,7 @@
 package cr_migration
 
 import (
-	miqv1alpha1 "github.com/ManageIQ/manageiq-pods/manageiq-operator/pkg/apis/manageiq/v1alpha1"
+	miqv1alpha1 "github.com/ManageIQ/manageiq-pods/manageiq-operator/api/v1alpha1"
 )
 
 func migrate20210503163000(cr *miqv1alpha1.ManageIQ) *miqv1alpha1.ManageIQ {
