+++ manageiq-operator/api/v1alpha1/helpers/miq-components/httpd_conf.go
@@ -2,7 +2,7 @@ package miqtools
 
 import (
 	"fmt"
-	miqv1alpha1 "github.com/ManageIQ/manageiq-pods/manageiq-operator/pkg/apis/manageiq/v1alpha1"
+	miqv1alpha1 "github.com/ManageIQ/manageiq-pods/manageiq-operator/api/v1alpha1"
 )
 
 // auth-configuration.conf
